bow and tf_idf creation (with problems)

+++ utility.cpp
@@ -44,7 +44,10 @@ void create_bow_and_tf(int** bow, float** tf_idf, Clique* list_of_entries, Bow* 
         // after finishing bow table
         // run through it and create tf values 
         for(int i=0 ; i<DICTIONARY_SIZE ; i++) {
-            tf_idf[entry_counter][i] = bow[entry_counter][i]/word_counter;
+            tf_idf[entry_counter][i] = (float)bow[entry_counter][i]/(float)word_counter;
+            // if(tf_idf[entry_counter][i] != 0) {
+            //     std::cout << tf_idf[entry_counter][i] << "\n";
+            // }
         }
 
         temp = temp->next;
